Switched to single precision for model fitting. Changed configure behaviour towards CFLAGS and LDFLAGS: automatic settings are now applied only if the --enable-automatic-flags configure option is set (thanks to S.Pascual). Version number pushed to 2.9.0.

+++ trunk/src/psf.h
@@ -68,7 +68,7 @@ typedef struct psf
   int		*masksize;	/* PSF mask dimensions */
   int		masknpix;	/* Total number of involved PSF pixels */
   float		*maskcomp;      /* Complete pix. data (PSF components) */
-  double	*maskloc;	/* Local PSF */
+  float		*maskloc;	/* Local PSF */
   double	**context;	/* Contexts */
   t_type	*contexttyp;	/* Context types */
   char		**contextname;	/* Array of context key-names */
